started refactoring pkg
