+++ tests/e2e/pkg/environment.go
@@ -1,3 +1,6 @@
+// Copyright (c) 2020, 2021, Oracle and/or its affiliates.
+// Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
+
 package pkg
 
 import (
