+++ tests/e2e/pkg/vmi.go
@@ -6,8 +6,6 @@ package pkg
 import (
 	"context"
 	"crypto/tls"
-	"encoding/json"
-	"errors"
 	"fmt"
 	"github.com/hashicorp/go-retryablehttp"
 	"net"
@@ -22,12 +20,6 @@ const (
 	RETRY_WAIT_MIN = 1 * time.Second
 	RETRY_WAIT_MAX = 30 * time.Second
 )
-
-type Elastic struct {
-	ClusterName   string `json:"cluster_name"`
-	binding       string
-	vmiHttpClient *retryablehttp.Client
-}
 
 func GetSystemVmiHttpClient() *retryablehttp.Client {
 	vmiRawClient := getHttpClientWIthCABundle(getSystemVMICACert())
@@ -58,47 +50,6 @@ func GetSystemVMICredentials() (*UsernamePassword, error) {
 	}, nil
 }
 
-//Gets Elastic representing the elasticsearch cluster with the binding name
-func GetElastic(binding string) *Elastic {
-	return &Elastic{
-		binding: binding,
-	}
-}
-
-//Checks if all elasticsearch required pods are running
-func (e *Elastic) PodsRunning() bool {
-	expectedElasticPods := []string{
-		fmt.Sprintf("vmi-%s-es-master", e.binding),
-		fmt.Sprintf("vmi-%s-kibana", e.binding),
-		fmt.Sprintf("vmi-%s-grafana", e.binding),
-		fmt.Sprintf("vmi-%s-prometheus", e.binding),
-		fmt.Sprintf("vmi-%s-api", e.binding)}
-	running := PodsRunning("verrazzano-system", expectedElasticPods)
-
-	if running {
-		expectedElasticPods = []string{
-			fmt.Sprintf("vmi-%s-es-ingest", e.binding),
-			fmt.Sprintf("vmi-%s-es-data", e.binding)}
-		running = PodsRunning("verrazzano-system", expectedElasticPods)
-	}
-
-	return running
-}
-
-//Checks if the elasticsearch cluster can be connected
-func (e *Elastic) Connect() bool {
-	esURL := GetVerrazzanoInstance().ElasticURL
-	body, err := e.retryGet(esURL, username, GetVerrazzanoPassword())
-	if err != nil {
-		return false
-	}
-	err = json.Unmarshal(body, e)
-	if err != nil {
-		return false
-	}
-	return true
-}
-
 func newRetryableHttpClient(client *http.Client) *retryablehttp.Client {
 	retryableClient := retryablehttp.NewClient() //default of 4 retries is sufficient for us
 	retryableClient.RetryMax = NUM_RETRIES
@@ -108,102 +59,11 @@ func newRetryableHttpClient(client *http.Client) *retryablehttp.Client {
 	return retryableClient
 }
 
-func (e *Elastic) retryGet(url, username, password string) ([]byte, error) {
-	req, _ := retryablehttp.NewRequest("GET", url, nil)
-	req.SetBasicAuth(username, password)
-	client := e.getVmiHttpClient()
-	client.CheckRetry = GetRetryPolicy(url)
-	resp, err := client.Do(req)
-	if err != nil {
-		Log(Info, fmt.Sprintf("Error GET %v error: %v", url, err))
-		return nil, err
-	}
-	if resp.StatusCode != 200 {
-		Log(Info, fmt.Sprintf("Response status code: %d", resp.StatusCode))
-	}
-	httpResp := ProcHttpResponse(resp, err)
-	if httpResp.StatusCode == http.StatusNotFound {
-		err = errors.New(fmt.Sprintf("url %s returned not found", url))
-		Log(Info, fmt.Sprintf("NotFound %v error: %v", url, err))
-		return nil, err
-	}
-	return httpResp.Body, nil
-}
-
-func (e *Elastic) getVmiHttpClient() *retryablehttp.Client {
-	if e.vmiHttpClient == nil {
-		e.vmiHttpClient = GetBindingVmiHttpClient(e.binding)
-	}
-	return e.vmiHttpClient
-}
-
-//Lists elasticsearch indices
-func (e *Elastic) ListIndices() []string {
-	idx := []string{}
-	for i, _ := range e.GetIndices() {
-		idx = append(idx, i)
-	}
-	return idx
-}
-
-//GetIndices gets index metadata (aliases, mappings, and settings) of all elasticsearch indices
-func (e *Elastic) GetIndices() map[string]interface{} {
-	esURL := GetVerrazzanoInstance().ElasticURL
-	body, err := e.retryGet(esURL, username, GetVerrazzanoPassword())
-	if err != nil {
-		Log(Info, fmt.Sprintf("Error ListIndices %v error: %v", esURL, err))
-		return nil
-	}
-	var indices map[string]interface{}
-	json.Unmarshal(body, &indices)
-	return indices
-}
 
 func GetBindingVmiHttpClient(bindingName string) *retryablehttp.Client {
 	bindingVmiCaCert := getBindingVMICACert(bindingName)
 	vmiRawClient := getHttpClientWIthCABundle(bindingVmiCaCert)
 	return newRetryableHttpClient(vmiRawClient)
-}
-
-//Lookup the Elasticsearch host
-func (e *Elastic) LookupHost() bool {
-	esURL := GetVerrazzanoInstance().ElasticURL
-	return Lookup(esURL)
-}
-
-//Check the Elasticsearch secret
-func (e *Elastic) CheckTlsSecret() bool {
-	secretName := fmt.Sprintf("%v-tls", e.binding)
-	return SecretsCreated("verrazzano-system", secretName)
-}
-
-//Check the Elasticsearch certificate
-func (e *Elastic) CheckCertificate() bool {
-	certList, _ := ListCertificates("verrazzano-system")
-	for _, cert := range certList.Items {
-		if cert.Name == fmt.Sprintf("%v-tls", e.binding) {
-			Log(Info, fmt.Sprintf("Found Certificate %v for binding %v", cert.Name, e.binding))
-			for _, condition := range cert.Status.Conditions {
-				if condition.Type == "Ready" {
-					Log(Info, fmt.Sprintf("Certificate %v status: Ready = %v", cert.Name, condition.Status))
-					return condition.Status == "True"
-				}
-			}
-		}
-	}
-	return false
-}
-
-//Check the Elasticsearch Ingress
-func (e *Elastic) CheckIngress() bool {
-	ingressList, _ := ListIngresses("verrazzano-system")
-	for _, ingress := range ingressList.Items {
-		if ingress.Name == fmt.Sprintf("vmi-%v-es-ingest", e.binding) {
-			Log(Info, fmt.Sprintf("Found Ingress %v for binding %v", ingress.Name, e.binding))
-			return true
-		}
-	}
-	return false
 }
 
 func getBindingVMICACert(bindingName string) []byte {
